Arrange All ViewComponents

+++ Shop_Example.Dtoes/Admin/Warranty/EditWarrantyDto.cs
@@ -10,7 +10,7 @@ namespace Shop_Example.Dtoes.Admin.Warranty
     public class EditWarrantyDto
     {
         [Required]
-        public long WarrantyId { get; set; }
+        public int WarrantyId { get; set; }
 
         [Required]
         [MaxLength(40, ErrorMessage = "نام گارانتی فقط میتواند 40 کاراکتر باشد")]
@@ -21,7 +21,7 @@ namespace Shop_Example.Dtoes.Admin.Warranty
         public string Description { get; set; }
 
         [Required]
-        public long ProductId { get; set; }
+        public int ProductId { get; set; }
 
     }
 }
